Rename index.js to module.js
Fix up import statements as well.
