+++ src/component/footer/footer.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import CacheButton from './cache';
-import Throttle from './throttle';
-import RequestCount from './request-count';
+import CacheButton from './cache.js';
+import Throttle from './throttle.js';
+import RequestCount from './request-count.js';
 
 const {func, object} = React.PropTypes;
