+++ src/component/title-bar/title-bar.js
